Mantenedor de Ofertas

+++ app/Http/Controllers/OfertasController.php
@@ -19,6 +19,88 @@ use App\Http\Requests\OfertaRequest;
 
 class OfertasController extends Controller
 {   
+
+    public function index2(Request $request)
+    {
+        $query = Oferta::query();
+
+
+        if (Gate::allows('empresa-gestion')) 
+        {
+            $emailUsuario = auth()->user()->correo_usuario; 
+            $empresaId = Empresa::where('id_usuario', $emailUsuario)->first()->id;
+    
+            // Filtrar ofertas solo de la empresa logueada
+            $query->where('id_empresa', $empresaId);
+        }
+        
+        // Filtrar por término de búsqueda
+        if ($request->filled('termino')) 
+        {
+            $query->where('titulo', 'LIKE', '%' . $request->input('termino') . '%');
+        }
+
+        // Filtrar por tipo
+        if ($request->filled('tipo')) 
+        {
+            $query->where('id_tipo', $request->input('tipo'));
+        }
+
+        // Filtrar por carrera
+        if ($request->filled('carrera')) 
+        {
+            $query->where('id_carrera', $request->input('carrera'));
+        }
+
+        // Filtrar por región
+        if ($request->filled('region')) 
+        {
+            $query->where('id_region', $request->input('region'));
+        }
+
+        // Filtrar por comuna
+        if ($request->filled('comuna')) 
+        {
+            $query->where('id_comuna', $request->input('comuna'));
+        }
+
+        // Filtro de la fecha 
+        if ($request->filled('rango_fecha'))
+        {
+            switch ($request->rango_fecha) {
+                case '1_semanas':
+                    $query->where('fecha_publicacion', '>=', Carbon::now()->subWeek());
+                    break;
+                case '2_semanas':
+                    $query->where('fecha_publicacion', '>=', Carbon::now()->subWeeks(2));
+                    break;
+                case '1_mes':
+                    $query->where('fecha_publicacion', '>=', Carbon::now()->subMonth());
+                    break;
+                case '2_meses':
+                    $query->where('fecha_publicacion', '>=', Carbon::now()->subMonths(2));
+                    break;
+                case '3_meses':
+                    $query->where('fecha_publicacion', '>=', Carbon::now()->subMonths(3));
+                    break;
+                case 'mas_3_meses': 
+                    $query->where('fecha_publicacion', '<', Carbon::now()->subMonths(3));
+                    break;
+            }
+        }
+
+        $ofertas = $query->get();
+
+        return view('ofertas.index2', 
+        [
+            'ofertas' => $ofertas,
+            'tipos' => Tipo::all(),
+            'carreras' => Carrera::all(),
+            'regiones' => Region::all(),
+            'comunas' => Comuna::all(),
+        ]);
+    }
+
     public function postular(Request $request)
     {
         $ofertaId = $request->oferta_id;
@@ -132,15 +214,30 @@ class OfertasController extends Controller
 
     public function destroy(Oferta $oferta)
     {
-        
-        $oferta->delete();
-        return redirect()->route('ofertas.index');
+        $cantPost = Postulacion::where('id_oferta', $oferta->id)->count();
+        if ($cantPost > 0) {
+            return back()->withErrors(['error'=>'No se puede eliminar la oferta porque tiene postulaciones']);
+        }else{
+            $oferta->delete();
+            if (Gate::allows('empresa-gestion')) {
+                return redirect()->route('ofertas.index');
+            } elseif (Gate::allows('admin-gestion')) {
+                return redirect()->route('ofertas.index2');
+            } else {
+                return redirect()->route('home');
+            }
+        }
     }
 
     public function create()
     {
-        $emailUsuario = auth()->user()->correo_usuario; 
-        $empresa = Empresa::where('id_usuario', $emailUsuario)->first();
+        if(Gate::allows('empresa-gestion'))
+        {
+            $emailUsuario = auth()->user()->correo_usuario; 
+            $empresa = Empresa::where('id_usuario', $emailUsuario)->first();
+        }else{
+            $empresa = Empresa::all();
+        }
         
 
         $regiones = Region::all();
@@ -163,22 +260,28 @@ class OfertasController extends Controller
         $oferta -> id_comuna = $request->comuna;
         $oferta -> id_carrera = $request->carrera;
         $oferta -> id_tipo = $request->tipo;
-
-        $emailUsuario = auth()->user()->correo_usuario; 
-        $empresa = Empresa::where('id_usuario', $emailUsuario)->first();
-
-        $oferta ->id_empresa = $empresa -> id;
-
+        $oferta ->id_empresa = $request->empresa;
 
         $oferta->save();
 
-        return redirect()->route('ofertas.index');
+        if (Gate::allows('empresa-gestion')) {
+            return redirect()->route('ofertas.index');
+        } elseif (Gate::allows('admin-gestion')) {
+            return redirect()->route('ofertas.index2');
+        } else {
+            return redirect()->route('home');
+        }
     }
 
     public function edit(Oferta $oferta)
     {
-        $emailUsuario = auth()->user()->correo_usuario; 
-        $empresa = Empresa::where('id_usuario', $emailUsuario)->first();
+        if(Gate::allows('empresa-gestion'))
+        {
+            $emailUsuario = auth()->user()->correo_usuario; 
+            $empresa = Empresa::where('id_usuario', $emailUsuario)->first();
+        }else{
+            $empresa = Empresa::all();
+        }
 
         $regiones = Region::all();
         $carreras = Carrera::all();
@@ -194,15 +297,22 @@ class OfertasController extends Controller
         $oferta -> cupos = $request->cupos;
         $oferta -> fecha_publicacion = now();
         $oferta -> descripcion = $request -> descripcion;
-
+        $oferta -> id_empresa = $request->empresa;
         $oferta -> id_region = $request->region;
         $oferta -> id_comuna = $request->comuna;
         $oferta -> id_carrera = $request->carrera;
         $oferta -> id_tipo = $request->tipo;
         
+        
         $oferta->save();
 
-        return redirect()->route('ofertas.index');
+        if (Gate::allows('empresa-gestion')) {
+            return redirect()->route('ofertas.index');
+        } elseif (Gate::allows('admin-gestion')) {
+            return redirect()->route('ofertas.index2');
+        } else {
+            return redirect()->route('home');
+        }
     }
 
     public function postulantes()
